Resolved issues commented
